refactor: 💡 space availability repository
This standards the storage of the Availability protobuf message.
Particularly when being called, it just simply wraps the leveldb
database. Functionally, this is required as a per day override may be
used in order to 'ram' the availability of a space type to 0 when a
guesthouse wants to 'block' out a specific space type for a day.

BREAKING CHANGE: 🧨 Discrete daily override does NOT only increment numSpaces